Refactor context card dashboard services
